move itemFilterService to Services/
